requirements.txt
clint==0.5.1
PyYAML==5.3.1
tqdm==4.56.0
wikiextractor